Rename /api to /apis
