+++ apis/mutations/newsletter/useSubmitNewsletterForm.ts
