+++ components/reusable-components/footer/Footer.tsx
@@ -17,7 +17,7 @@ import Button from '../button/Button';
 import {
   NewsletterFormData,
   useSubmitNewsletterForm,
-} from '../../../api/mutations/newsletter/useSubmitNewsletterForm';
+} from '../../../apis/mutations/newsletter/useSubmitNewsletterForm';
 
 // eslint-disable-next-line no-unused-vars
 interface FormValues {
